Permissions changes in move-space
Previously we permissions checked move-space syncs with the to_space in
both the delete and the add call. now, when moving a note/board between
spaces, we look up that item's space_id (directly from db since it's
plaintext) and use it in the `delete` check. this was a lucky find while
doing a smaug-like gold-fondling of the code on github the other day. i
guess it helps to be obsessive about code...this but probably wouldn't
have ever caused too much of an issue since it will be enforced by the
interface anyway, but it's good to keep things in check.