Locking updates!
Converting all locks to use new macros:

x.lock().unwrap() -> lock!(x)
x.read().unwrap() -> lockr!(x)
x.write().unwrap() -> lockw!(x)

the purpose is to give a central place to control all locking. if we
want to add deadlock debugging to the app, it can be done in one place
now as opposed to ~170 places.

not only this, but if we decide later to use a different locking scheme
or add more checks, we can easily update the macros. great.

!!!NOTE!!! found some places where we were using db.read().unwrap()...
this is really bad because the db must be write locked (its in a TODO to
convert db from RwLock to Mutex). so this change saved us a few really
obscure panics!

all tests passing.