Convert turtl.db from RwLock to Mutex
The old code was dangerous because running more than one query at a time
on the turtl.db object (which is also shared across all our syncers)
will panic due to the internal nature of the sqlite crate we use.

RwLock.read() allows multiple queries to run at once, which could easily
have caused uncaught panics. the old solution was to use db.write() to
lock by convention, but after converting all locking to use the new
macros, i realized convention is a terrible (and inaccurate) way to
handle this.

all references to the user's db (turtl.db and its copies in the sync
system) have been converted to use Mutex instead of RwLock, which
removes all ambiguity from the locking.

tests passing. TODO updated (there was an item for this).