import/export done!
...integration tests as well. WOW.

one caveat: if a user imports an export from another user (or server),
all the models from it will be created, and their ids renamed. if that
same export is used to import again, the models in it will be DUPLICATED
because the ids will not have anything to compare against.

so either models get duplicated and users can move data between servers,
or models are NOT duplicated and an import can only be imported a) by
the same users or b) on a different server.