Solved the import problem!
we're now using a deterministic client_id when importing. it's a bit
weird to do things this way, but for the imported items, any ids that
dont have a matching record int he db get assigned the same id with the
client_id swapped out (so timestamp/counter val are the same as before).
the new cid is a deterministic value -- hash of user_id and server. this
means that when you import the same export twice on a new user/new
server, it no longer double-adds the items. i just couldnt live with the
thought of this, and this was the solution.