Big commit, nothing specific
- adding `wrap_errors` config option that unwraps errors if returning
them to the client
- changing the way errors are displayed. basically, making them all json
now. the idea is that when errors are returned to the ui, they are a) a
json object that can be easily parsed and b) a standard format
(somewhat)
- adding username/password validation to user::join (instead of relying
on the ui to do this). the only issue is that the i18n stuff is now
getting muddier.
- ripping out event emitter from everywhere. turns out it was only being
used a miniscule amount of places, and really just kind of made things
more convoluted.
- making all integration tests wait for "user:login" event after calling
login/join so they know the user is ready to use (the event may fire
after the dispatch returns).
- adding poll before sync_start to make sure turtl.db exists