outgoing file sync fixes
- when grabbing the next outgoing file, check if its sync item is frozen
ebfore proceeding! we were not checking before, so failures were just
looping endlessly.
- setting file upload timeout to 60s (up from the default, which was
probably too low)
- making it so if we try to find a file for upload and it doesn't exist,
we include that failure in the "errors that count against the sync items
error count" list so that if it happens more than 3 times, we freeze the
sync item

this fixes turtl/tracker#176. yay!!

note that it doesn't actually stop the race condition, but it does stop
the core from going haywire when it happens. the race condition itself
is somewhat of an edge case at this point, and if the core reports the
error to the "Sync info" page (which it now does) then i am happy.